==> revamp/10.2_phys-heatmaps.Rmd <==
---
title: "Untitled"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

#Setup 

```{r}

"%&%" <- function(a,b) paste0(a,b)

library("tidyverse")
library("data.table")
library("Homo.sapiens")

fuse.path <- "/home/jason/science/servers/"
serv.dir <- fuse.path %&% "FUSE5/"
proj.dir <- serv.dir %&% "projects/t2d_classification/"
methA.dir <- proj.dir %&% "method_A/"
work.dir <- proj.dir %&% "revamp/"
analysis.dir <- work.dir %&% "analysis_files/"
plot.dir <- work.dir %&% "plots/"


cred.df <- fread(work.dir %&% "genetic_credible_sets/gencred.txt")

class.df1 <- fread(analysis.dir %&% "classified-loci_weighted.txt")
class.df2 <- fread(analysis.dir %&% "classified-loci_weighted_with-shared.txt")

```


```{r}

make_maxppa_df <- function(){
  id.vec <- cred.df$CondID %>% unique(.)
  pb <- txtProgressBar(min=0,max=length(id.vec),style=3)
  ppa.vec <- c() 
  for (i in 1:length(id.vec)){
    setTxtProgressBar(pb,i)
    id <- id.vec[i]
    ppa <- (filter(cred.df,CondID==id) %>% arrange(.,desc(PPA)))$PPA[1]
    ppa.vec <- append(ppa.vec,ppa)
  } 
  out.df <- data.frame(Locus.ID=id.vec,maxppa=ppa.vec,stringsAsFactors = FALSE)
  return(out.df)
}


```






```{r}

library("plotly")
library("heatmaply")

maxppa.df <- make_maxppa_df()
weighted.df <- fread(analysis.dir %&% "tissue_ppa_divvy-full-weighted-unscaled.txt")
w.df <- weighted.df
phys94.df <- fread(work.dir %&% "analysis_files/phys94-ld-assigned.txt")
physfull.df <- inner_join(phys94.df,w.df,by="Locus.ID") %>% filter(.,assigned.r2>=0.3,other<=0.50)
mat <- dplyr::select(physfull.df,one_of("islet","adipose","muscle",
                                        "liver")) %>%
  as.matrix(.)

name.vec <- map(1:dim(physfull.df)[1],function(i){
  sub <- physfull.df[i,]
  p <- sub$Physiology
  p <- ifelse(p=="IA","Insulin-Action",p)
  p <- ifelse(p=="IS-1","Insulin-Secretion-1",p)
  p <- ifelse(p=="IS-2","Insulin-Secretion-2",p)
  p %&% "_" %&% sub$symbol
}) %>% as.character(.)

row.names(mat) <- name.vec
mat.df <- as.data.frame(mat)


```


```{r}

#heatmaply((mat.df), k_row = 1, k_col = 1,fontsize_row=3,plot_method = "ggplot")
```




```{r}

physfull.df <- arrange(physfull.df,Physiology)

```

```{r}

library("plyr")
library("scales")

```


```{r}

physfull.df <- inner_join(phys94.df,w.df,by="Locus.ID") %>% filter(.,assigned.r2>=0.3,other<=0.50)

plot.df <- dplyr::select(physfull.df,
             one_of("Physiology","symbol","islet","adipose",
                                "muscle","liver"))
plot.df$Name <-  map(1:dim(plot.df)[1],function(i){
  sub <- plot.df[i,]
  p <- sub$Physiology
  p <- ifelse(p=="IA","Insulin-Action",p)
  p <- ifelse(p=="IS-1","Insulin-Secretion-1",p)
  p <- ifelse(p=="IS-2","Insulin-Secretion-2",p)
  p %&% " (" %&% sub$symbol %&% ")"
}) %>% as.character(.)

plot.df <- dplyr::select(plot.df,-one_of("Physiology","symbol"))

plot.df$Name <- as.factor(plot.df$Name)
plot.df <- melt(plot.df)

plot.df$Name <- factor(plot.df$Name,levels=c("Adiposity (FTO)",
                                             "Adiposity (NFAT5)",
                     "Adiposity (POC5)","Dyslipidemia (GCKR)",
                     "Dyslipidemia (TM6SF2)",
                     "Dyslipidemia (TOMM40-APOE)",
                     "Insulin-Action (ANKRD55)",
                     "Insulin-Action (BPTF)","Insulin-Action (CMIP)",
                     "Insulin-Action (FAM13A)","Insulin-Action (GRB14-COBLL1)","Insulin-Action (HNF4A)","Insulin-Action (KIF9)",
                     "Insulin-Action (MACF1)","Insulin-Action (PEPD)",
                     "Insulin-Action (PPARG)","Insulin-Action (VEGFA)",
                     "IA-IS-combination (CEP68)",
                     "IA-IS-combination (FAM63A)","IA-IS-combination (GIPR)","IA-IS-combination (HNF1A)",
                     "IA-IS-combination (HNF1B)",
                     "IA-IS-combination (HNF4A)",
                     "IA-IS-combination (MLX)",
                     "IA-IS-combination (MPHOSPH9)",
                     "IA-IS-combination (MTMR3-ASCC2)",
                     "IA-IS-combination (PAM)","IA-IS-combination (PLEKHA1)","IA-IS-combination (PNPLA3)","IA-IS-combination (PRC1)",
                     "IA-IS-combination (PTPN9)",
                     "IA-IS-combination (RREB1)",
                     "IA-IS-combination (SPRY2)",
                     "IA-IS-combination (TLE1)",
                     "IA-IS-combination (TMEM154)",
                     "IA-IS-combination (WFS1)",
                     "IA-IS-combination (ZMIZ1)",
                     "IA-IS-combination (ZZEF1)",
                     "Insulin-Secretion-2 (ADAMTS9)",
                     "Insulin-Secretion-2 (ANK1)",
                     "Insulin-Secretion-2 (ANKH)",
                     "Insulin-Secretion-2 (C2CD4A-B)",
                     "Insulin-Secretion-2 (CCND2)",
                     "Insulin-Secretion-2 (CDKAL1)",
                     "Insulin-Secretion-2 (CDKN2A-B)",
                     "Insulin-Secretion-2 (CENTD2-ARAP1)",
                     "Insulin-Secretion-2 (GLIS3)",
                     "Insulin-Secretion-2 (GPSM1)",
                     "Insulin-Secretion-2 (HHEX-IDE)",
                     "Insulin-Secretion-2 (HMG20A)",
                     "Insulin-Secretion-2 (JAZF1)",
                     "Insulin-Secretion-2 (KCNQ1)",
                     "Insulin-Secretion-2 (KLHDC5)",
                     "Insulin-Secretion-2 (PROX1)",
                     "Insulin-Secretion-2 (RREB1)",
                     "Insulin-Secretion-2 (THADA)",
                     "Insulin-Secretion-2 (ZBED3)",
                     "Insulin-Secretion-1 (ADCY5)",
                     "Insulin-Secretion-1 (GCK)",
                     "Insulin-Secretion-1 (MTNR1B)",
                     "Insulin-Secretion-1 (SLC30A8)",
                     "Insulin-Secretion-1 (TCF7L2)"))

#colors=c(viridis(100)[10:50],viridis(100)[60:100])
p <- ggplot(plot.df, aes(variable, Name)) + 
  geom_tile(aes(fill = value),colour = "black") + 
  scale_fill_gradientn(colors=c(viridis(100)[1:15],
                                viridis(100)[30:45],
                                viridis(100)[60:75],
                                viridis(100)[85:100])) + 
  theme(axis.text.x = element_text(size=9,angle=30,
                                   vjust = 0.5),
        axis.text.y = element_text(size=6,
                                   color=c(rep(viridis(10)[3],3),
                                           rep("black",3),
                                           rep(viridis(10)[3],11),
                                           rep("black",21),
                                           rep(viridis(10)[3],19),
                                           rep("black",6)))) + 
  xlab("Tissue-of-action Score") 

ggsave(plot=p,filename = plot.dir %&% "phys-tile.png",
       height=6,width=5)
ggsave(plot=p,filename = plot.dir %&% "phys-tile.pdf",
       height=6,width=5)


```



```{r}

phys.df2 <- fread(work.dir %&% "analysis_files/physUDLER-ld-assigned.txt")
physfull.df2 <- inner_join(phys.df2,w.df,by="Locus.ID") %>% filter(.,assigned.r2>=0.3,other<=0.50)

plot.df2 <- dplyr::select(physfull.df2,
             one_of("Physiology","symbol","islet","adipose",
                                "muscle","liver"))
plot.df2$Name <-  map(1:dim(plot.df2)[1],function(i){
  sub <- plot.df2[i,]
  p <- sub$Physiology
  p %&% " (" %&% sub$symbol %&% ")"
}) %>% as.character(.)

plot.df2 <- dplyr::select(plot.df2,-one_of("Physiology","symbol"))

plot.df2$Name <- as.factor(plot.df2$Name)
plot.df2 <- melt(plot.df2)
p2 <- ggplot(plot.df2, aes(variable, Name)) + 
  geom_tile(aes(fill = value),colour = "ghostwhite") + 
  scale_fill_gradientn(colors=c(viridis(100)[10:20],
                       viridis(100)[30:100])) + 
  theme(axis.text.x = element_text(size=9,angle=30,
                                   vjust = 0.5),
        axis.text.y = element_text(size=6,
                                   color=c(rep(viridis(10)[3],23),
                                           rep("black",18),
                                           rep(viridis(10)[3],3),
                                           rep("black",5),
                                           rep(viridis(10)[3],4)))) + 
  xlab("Tissue-of-action Score") 

ggsave(plot=p2,filename = plot.dir %&% "phys-tile_Udler.png",
       height=6,width=5)
```
