==> revamp/15.1_gwas-enrich-input.Rmd <==
---
title: "14.1_eQTL-enrich-revamp-make-inputs.Rmd"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# Setup

```{r}

"%&%" <- function(a,b) paste0(a,b)

library("tidyverse")
library("data.table")

serv.dir1 <- "/home/jason/science/servers/FUSE/" 
serv.dir2 <- "/home/jason/science/servers/FUSE5/" 
proj.dir <- serv.dir2 %&% "projects/t2d_classification/"
work.dir <- proj.dir %&% "revamp/"
out.dir <- work.dir %&% "enrichment_files/gwas/"

cred.df <- fread(work.dir %&% "genetic_credible_sets/gencred.txt")

```

create and save signal file 

```{r}

sub.df <- dplyr::select(cred.df,one_of("symbol","IndexSNP","lead.rsid","CondID"))
sub.df <- sub.df[!duplicated(sub.df),]
write.table(x=sub.df$lead.rsid,file = work.dir%&%"analysis_files/ensembl_query_snps.txt",
            sep="\t",quote=F,row.names=F,col.names=F)

```


Manually uploaded file of rsids to e!Ensembl for SNP query of SNP coordinates in Build GRch 38) (accessed Aug 26, 2018)

NOTE: SNP rs505922 (GRCh37, chr9:136149229) at the ABO locus did not yield a result from the Ensembl query, the remaining 379 SNPs did 


```{r}

ens.df <- fread(work.dir %&% "analysis_files/ensembl_query_results.txt")
names(ens.df) <- c("rsid","source","chrom","bp.start","bp.end")
sub.df$grch38 <- map(1:dim(sub.df)[1],function(i){
  row.df <- sub.df[i,]
  chromo <- (strsplit(row.df$IndexSNP,split=":")[[1]][1] %>% strsplit(.,split="chr"))[[1]][2]
  r <- row.df$lead.rsid
  sub <- filter(ens.df,rsid==r,chrom==chromo)
  snp38 <- sub$chrom%&%":"%&%sub$bp.start
}) %>% as.character(.)
write.table(x=sub.df,file = work.dir%&%"analysis_files/index_snps.txt",
            sep="\t",quote=F,row.names=F)

```





Process GWAS input (downloaded from NHGRI-EBI catalogue accessed Aug 23, 2019)

```{r}

gwas.df <- fread(serv.dir2 %&%
                   "datasets/nhgri_ebi_gwas_catalogue/gwas_catalog_v1.0-associations_e96_r2019-07-30.tsv",sep="\t")

sub.df <- dplyr::select(gwas.df,one_of("DISEASE/TRAIT","CHR_ID","CHR_POS"))
sub.df$SNPID <- "chr"%&%sub.df$CHR_ID %&% ":" %&% sub.df$CHR_POS
sub.df <- dplyr::select(sub.df,one_of("SNPID","DISEASE/TRAIT"))
names(sub.df) <- c("snp","pheno")

write.table(x=sub.df,file = out.dir %&% "catalogue.txt",
            sep="\t",quote=F,row.names=F)

```
